Fix how longs are created

+++ src/main/java/org/imdea/fixcheck/transform/input/provider/LongProvider.java
@@ -12,7 +12,7 @@ public class LongProvider implements InputProvider {
 
   @Override
   public Object getInput() {
-    return random.nextLong();
+    return (long) random.nextInt();
   }
 
   @Override
